requirements.txt
flask
flask_sqlalchemy
markdown
passlib
validate_email
BeautifulSoup4
html2text
pymysql
requests
pillow
nose
blinker